Feat : 게시글과 댓글 간의 일대다 연결

+++ src/main/java/com/example/minipro2_post/repository/CommentRepository.java
@@ -3,5 +3,6 @@ package com.example.minipro2_post.repository;
 import com.example.minipro2_post.entity.CommentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
+
+public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
 }
